+++ backend/src/Logitar.Portal.GraphQL/AggregateGraphType.cs
@@ -8,7 +8,7 @@ internal abstract class AggregateGraphType<T> : ObjectGraphType<T> where T : Agg
 {
   protected AggregateGraphType(string? description = null)
   {
-    Name = typeof(T).Name;
+    Name = typeof(T).Name.Remove("Model");
     Description = description;
 
     Field(x => x.Id)
